// src/app/enums/local-storage-keys.enums.ts
export enum LocalStorageKeys {
  ACTA_STORAGE_KEY = 'actas_storage_key',
  AVAILABLE_ACTAS = 'available_actas',
  CALLES = 'calles',
  COLONIAS = 'colonias',
  CIUDADES = 'ciudades',
  LOCATION_STORAGE_KEY = 'location_storage_key',
  MULTIPLE_OPTIONS = 'multiple_options',
  REQUEST_STORAGE_KEY = 'requeststoragekey',
  PENDING_TO_SEND_ACTAS = 'pending_to_send_actas',
  PENDING_REQUEST = 'pending_request',
  PENDING_REQUEST_ACTA = 'pending_request_acta',
  PENDING_REQUEST_MUL_OPT = 'pending_request_mul_opt',
  PENDING_REQUEST_IMAGES = 'pending_request_images',
  USER_KEY = 'user_storage_key',
  USER_INFO = 'user_info'
}